requirements.txt
alembic==1.11.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boto3==1.21.22
botocore==1.24.46
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
comm==0.1.3
databricks-cli==0.17.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
docker==6.1.2
email-validator==1.3.1
entrypoints==0.4
executing==1.2.0
fastapi==0.86.0
fastjsonschema==2.17.1
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonschema==4.17.3
jupyter-server==1.24.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab==3.3.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
mlflow==1.26.1
mysql-connector-python==8.0.33
nbclassic==0.5.6
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
orjson==3.8.12
packaging==23.1
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
prometheus-client==0.17.0
prometheus-flask-exporter==0.22.4
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
PyJWT==2.7.0
PyMySQL==1.0.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
querystring-parser==1.2.4
requests==2.31.0
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.20.4
tabulate==0.9.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.0
ujson==5.7.0
urllib3==1.26.16
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.3.4
xgboost==1.7.5
zipp==3.15.0
